feat: 콘솔 입력 기능 및 레이싱 게임 로직 분리
- 콘솔 입력 기능을 InputUtils.kt로 이동
- 레이싱 게임 로직을 RacingGame.kt로 이동
