+++ src/main/kotlin/racingcar/Application.kt
@@ -1,80 +1,10 @@
 package racingcar
 
-import camp.nextstep.edu.missionutils.Console
-import camp.nextstep.edu.missionutils.Randoms
-import kotlin.IllegalArgumentException
-
 fun main() {
-    println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분")
-
-    val names = try {
-        Console.readLine()
-    } catch (e: Exception) {
-        ""
-    }
-
-    println("시도할 횟수는 몇 회인가요?")
-
-    val tryNumber = try {
-        Console.readLine()
-    } catch (e: Exception) {
-        ""
-    }
+    val names = readCarNames()
+    val tryNumber = readTryNumber()
 
     println("실행 결과")
 
     winner(names, tryNumber.toInt())
-}
-
-fun validateName(name: String): String {
-    if (name.length > 5) {
-        throw IllegalArgumentException("이름은 5자 이하여야 합니다.")
-    }
-    return name
-}
-
-fun divideNames(names: String): List<String> {
-    val delimiter = ","
-
-    return names.split(delimiter.toRegex()).map { name ->
-        validateName(name)
-    }
-}
-
-fun randomMoveForward(name: String, position: MutableMap<String, Int>) {
-    val randomValue = Randoms.pickNumberInRange(0, 9)
-    if (randomValue >= 4) {
-        position[name] = position[name]!! + 1
-    }
-}
-
-fun currentPosition(names: String, position: Map<String, Int>) {
-    val finalParticipants = divideNames(names)
-
-    finalParticipants.forEach { name ->
-        val currentPosition = "-".repeat(position[name]!!)
-        println("$name: $currentPosition")
-    }
-}
-
-fun race(names: String, tryNumber: Int): Map<String, Int> {
-    val finalParticipants = divideNames(names)
-    val position = finalParticipants.associateWith { 0 }.toMutableMap()
-
-    for (trying in 1..tryNumber) {
-        finalParticipants.forEach { name ->
-            randomMoveForward(name, position)
-        }
-        currentPosition(names, position)
-        print("\n")
-    }
-    return position
-}
-
-fun winner(names: String, tryNumber: Int) {
-    val position = race(names, tryNumber)
-    val maxPosition = position.values.maxOrNull()
-    val winners = position.filter {it.value == maxPosition}.keys
-
-    println("최종 우승자 : ${winners.joinToString(", ")}")
 }
