+++ src/main/kotlin/racingcar/InputUtils.kt
@@ -1,0 +1,21 @@
+package racingcar
+
+import camp.nextstep.edu.missionutils.Console
+
+fun readCarNames(): String {
+    println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)")
+    return try {
+        Console.readLine() ?: ""
+    } catch (e: Exception) {
+        ""
+    }
+}
+
+fun readTryNumber(): String {
+    println("시도할 횟수는 몇 회인가요?")
+    return try {
+        Console.readLine() ?: ""
+    } catch (e: Exception) {
+        ""
+    }
+}
